numa_avatar/requirements.txt
fastapi
httpx
pyyaml